% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fc.R
\name{calc_fc}
\alias{calc_fc}
\title{Calculate fold change from observed and expected depth-of-coverage}
\usage{
calc_fc(obs, exp, dir)
}
\arguments{
\item{obs}{Observed depth-of-coverage}

\item{exp}{Expected depth-of-coverage}

\item{dir}{The direction of fold change to calculate - "gain" or "loss"}
}
\value{
The calculated fold change
}
\description{
Calculate fold change from observed and expected depth-of-coverage
}
\examples{
calc_fc(10, 2, dir = "gain")
}
