% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2r.R
\name{calc_log2r}
\alias{calc_log2r}
\title{Calculate log2 ratios}
\usage{
calc_log2r(obs, exp, digits = 3)
}
\arguments{
\item{obs}{Observed reads}

\item{exp}{Expected reads}

\item{digits}{Number of decimal places}
}
\value{
Log2 ratio
}
\description{
Calculate log2 ratios
}
\examples{
calc_log2r(10, 5)
}
