% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tc.R
\name{calc_tc}
\alias{calc_tc}
\title{Calculate total copies from copies per cell}
\usage{
calc_tc(cpc, cells = 100)
}
\arguments{
\item{cpc}{Copies per cell}

\item{cells}{Total number of cells. Defaults to 100.}
}
\value{
Total copies of a gene locus
}
\description{
Calculate total copies from copies per cell
}
\examples{
calc_tc(10, 100)
}
