% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Change between fold change and log2 ratios}
\usage{
change(input, from, to)
}
\arguments{
\item{input}{Input value}

\item{from}{The format to change from}

\item{to}{The format to change to}
}
\value{
The calculated output
}
\description{
Change between fold change and log2 ratios
}
\examples{
change(-1, from = "log2r", to = "fcloss")
}
