% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{Check fold change direction argument is appropriate}
\usage{
check_dir(x)
}
\arguments{
\item{x}{Fold change direction}
}
\value{
Errors if fold change not supplied as "gain" or "loss"
}
\description{
Check fold change direction argument is appropriate
}
\examples{
check_dir("loss")
}
