% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ncc.R
\name{check_ncc}
\alias{check_ncc}
\title{Check neoplastic cell content is in correct format}
\usage{
check_ncc(ncc)
}
\arguments{
\item{ncc}{Neplastic cell content in percentage format (10\% not 0.1)}
}
\value{
Errors if incorrect format supplied
}
\description{
Check neoplastic cell content is in correct format
}
\examples{
check_ncc(100)
}
