% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_cpc.R
\name{pred_cpc}
\alias{pred_cpc}
\title{Predict copies per cell}
\usage{
pred_cpc(fc, ncc, dir, ref_cpc = 2, stroma_cpc = 2, digits = 0)
}
\arguments{
\item{fc}{Fold change}

\item{ncc}{Neoplastic cell content (0-100)}

\item{dir}{Fold change direction, must be "gain" or "loss"}

\item{ref_cpc}{Copies per cell for reference cohort}

\item{stroma_cpc}{Copies per cell for patient stroma}

\item{digits}{The number of decimal places to round to}
}
\value{
Copies per cell
}
\description{
Predict copies per cell
}
