% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fc.R
\name{pred_fc}
\alias{pred_fc}
\title{Predict fold change from copies per cell}
\usage{
pred_fc(tumour_cpc, ncc, dir, ref_cpc = 2, stroma_cpc = 2, digits = 1)
}
\arguments{
\item{tumour_cpc}{Copies of gene locus per tumour cell}

\item{ncc}{Neoplastic cell content}

\item{dir}{Fold change direction, must be "gain" or "loss"}

\item{ref_cpc}{Copies of gene locus per reference cohort cell}

\item{stroma_cpc}{Copies of gene locus per stroma cell}

\item{digits}{Digits to round the fold change to}
}
\description{
Predict fold change from copies per cell
}
